fix: change typescript to peerDependency
When a consumer of this package has a different version of `typescript` (5.8.2, but this package uses 5.0.4) it causes 2 versions to be present, and war3-transformer will try to parse AST from a newer `typescript` using an older version.

Moving `typescript` to peerDependency will result in this package using consumer's dependency instead and avoiding having duplicate dependencies. This approach is used by `tsutils` and `typescript-to-lua`.